rename src/components to src/aframe-components
