+++ src/aframe-components/blending-opacity.js
