+++ src/aframe-components/create-from-json.js
