+++ src/aframe-components/css2d-renderer.js
