+++ src/aframe-components/focus-animation.js
