+++ src/aframe-components/gltf-part.js
