+++ src/aframe-components/google-maps-aerial.js
