+++ src/aframe-components/intersection.js
