+++ src/aframe-components/managed-street.js
