+++ src/aframe-components/notify.js
