+++ src/aframe-components/obb-clipping.js
