+++ src/aframe-components/polygon-offset.js
