+++ src/aframe-components/screentock.js
