+++ src/aframe-components/street-align.js
