+++ src/aframe-components/street-environment.js
