+++ src/aframe-components/street-generated-clones.js
