+++ src/aframe-components/street-generated-pedestrians.js
