+++ src/aframe-components/street-generated-rail.js
