+++ src/aframe-components/street-generated-stencil.js
