+++ src/aframe-components/street-generated-striping.js
