+++ src/aframe-components/street-geo.js
