+++ src/aframe-components/street-ground.js
