+++ src/aframe-components/street-label.js
