+++ src/aframe-components/street-mapping-streetplan.js
