+++ src/aframe-components/street-segment.js
