+++ src/aframe-components/svg-extruder.js
