+++ src/index.js
@@ -8,36 +8,36 @@ STREET.utils.newScene = streetUtils.newScene;
 var streetmixParsers = require('./aframe-streetmix-parsers');
 var streetmixUtils = require('./tested/streetmix-utils');
 
-require('./components/gltf-part');
-require('./components/ocean');
-require('./components/svg-extruder.js');
+require('./aframe-components/gltf-part');
+require('./aframe-components/ocean');
+require('./aframe-components/svg-extruder.js');
 require('./lib/animation-mixer.js');
 require('./lib/aframe-gaussian-splatting-component.min.js');
 require('./assets.js');
-require('./components/notify.js');
-require('./components/create-from-json');
-require('./components/screentock.js');
-require('./components/focus-animation');
+require('./aframe-components/notify.js');
+require('./aframe-components/create-from-json');
+require('./aframe-components/screentock.js');
+require('./aframe-components/focus-animation');
 require('aframe-atlas-uvs-component');
-require('./components/street-geo.js');
-require('./components/street-environment.js');
-require('./components/intersection.js');
-require('./components/obb-clipping.js');
-require('./components/street-segment.js');
-require('./components/managed-street.js');
-require('./components/street-generated-stencil.js');
-require('./components/street-generated-striping.js');
-require('./components/street-generated-pedestrians.js');
-require('./components/street-generated-rail.js');
-require('./components/street-generated-clones.js');
-require('./components/polygon-offset.js');
-require('./components/street-align.js');
-require('./components/street-ground.js');
-require('./components/street-label.js');
-require('./components/blending-opacity.js');
-require('./components/measure-line.js');
-require('./components/css2d-renderer.js');
-require('./components/google-maps-aerial.js');
+require('./aframe-components/street-geo.js');
+require('./aframe-components/street-environment.js');
+require('./aframe-components/intersection.js');
+require('./aframe-components/obb-clipping.js');
+require('./aframe-components/street-segment.js');
+require('./aframe-components/managed-street.js');
+require('./aframe-components/street-generated-stencil.js');
+require('./aframe-components/street-generated-striping.js');
+require('./aframe-components/street-generated-pedestrians.js');
+require('./aframe-components/street-generated-rail.js');
+require('./aframe-components/street-generated-clones.js');
+require('./aframe-components/polygon-offset.js');
+require('./aframe-components/street-align.js');
+require('./aframe-components/street-ground.js');
+require('./aframe-components/street-label.js');
+require('./aframe-components/blending-opacity.js');
+require('./aframe-components/measure-line.js');
+require('./aframe-components/css2d-renderer.js');
+require('./aframe-components/google-maps-aerial.js');
 require('./editor/index.js');
 var firebase = require('./editor/services/firebase.js');
